expose AttrDefGenerator and TypeDefGenerator
