+++ mlir/include/mlir/TableGen/AttrOrTypeDef.h
@@ -16,6 +16,7 @@
 
 #include "mlir/Support/LLVM.h"
 #include "mlir/TableGen/Builder.h"
+#include "mlir/TableGen/Class.h"
 #include "mlir/TableGen/Constraint.h"
 #include "mlir/TableGen/Trait.h"
 
@@ -27,6 +28,9 @@ class SMLoc;
 
 namespace mlir {
 namespace tblgen {
+class MethodParameter;
+class MethodParameter;
+class InterfaceMethod;
 class Dialect;
 
 //===----------------------------------------------------------------------===//
@@ -284,6 +288,148 @@ public:
   /// Get the unique type name "dialect.typename".
   StringRef getTypeName() const;
 };
+
+class DefGen {
+public:
+  /// Create the attribute or type class.
+  DefGen(const AttrOrTypeDef &def);
+
+  void emitDecl(raw_ostream &os) const;
+  void emitDef(raw_ostream &os) const;
+
+private:
+  /// Add traits from the TableGen definition to the class.
+  void createParentWithTraits();
+  /// Emit top-level declarations: using declarations and any extra class
+  /// declarations.
+  void emitTopLevelDeclarations();
+  /// Emit the function that returns the type or attribute name.
+  void emitName();
+  /// Emit the dialect name as a static member variable.
+  void emitDialectName();
+  /// Emit attribute or type builders.
+  void emitBuilders();
+  /// Emit a verifier declaration for custom verification (impl. provided by
+  /// the users).
+  void emitVerifierDecl();
+  /// Emit a verifier that checks type constraints.
+  void emitInvariantsVerifierImpl();
+  /// Emit an entry poiunt for verification that calls the invariants and
+  /// custom verifier.
+  void emitInvariantsVerifier(bool hasImpl, bool hasCustomVerifier);
+  /// Emit parsers and printers.
+  void emitParserPrinter();
+  /// Emit parameter accessors, if required.
+  void emitAccessors();
+  /// Emit interface methods.
+  void emitInterfaceMethods();
+
+  //===--------------------------------------------------------------------===//
+  // Builder Emission
+
+  /// Emit the default builder `Attribute::get`
+  void emitDefaultBuilder();
+  /// Emit the checked builder `Attribute::getChecked`
+  void emitCheckedBuilder();
+  /// Emit a custom builder.
+  void emitCustomBuilder(const AttrOrTypeBuilder &builder);
+  /// Emit a checked custom builder.
+  void emitCheckedCustomBuilder(const AttrOrTypeBuilder &builder);
+
+  //===--------------------------------------------------------------------===//
+  // Interface Method Emission
+
+  /// Emit methods for a trait.
+  void emitTraitMethods(const InterfaceTrait &trait);
+  /// Emit a trait method.
+  void emitTraitMethod(const InterfaceMethod &method);
+
+  //===--------------------------------------------------------------------===//
+  // Storage Class Emission
+  void emitStorageClass();
+  /// Generate the storage class constructor.
+  void emitStorageConstructor();
+  /// Emit the key type `KeyTy`.
+  void emitKeyType();
+  /// Emit the equality comparison operator.
+  void emitEquals();
+  /// Emit the key hash function.
+  void emitHashKey();
+  /// Emit the function to construct the storage class.
+  void emitConstruct();
+
+  //===--------------------------------------------------------------------===//
+  // Utility Function Declarations
+
+  /// Get the method parameters for a def builder, where the first several
+  /// parameters may be different.
+  SmallVector<MethodParameter>
+  getBuilderParams(std::initializer_list<MethodParameter> prefix) const;
+
+  //===--------------------------------------------------------------------===//
+  // Class fields
+
+  /// The attribute or type definition.
+  const AttrOrTypeDef &def;
+  /// The list of attribute or type parameters.
+  ArrayRef<AttrOrTypeParameter> params;
+  /// The attribute or type class.
+  Class defCls;
+  /// An optional attribute or type storage class. The storage class will
+  /// exist if and only if the def has more than zero parameters.
+  std::optional<Class> storageCls;
+
+  /// The C++ base value of the def, either "Attribute" or "Type".
+  StringRef valueType;
+  /// The prefix/suffix of the TableGen def name, either "Attr" or "Type".
+  StringRef defType;
+};
+
+class DefGenerator {
+public:
+  bool emitDecls(StringRef selectedDialect);
+  bool emitDefs(StringRef selectedDialect);
+
+protected:
+  DefGenerator(ArrayRef<const llvm::Record *> defs, raw_ostream &os,
+               StringRef defType, StringRef valueType, bool isAttrGenerator);
+
+  /// Emit the list of def type names.
+  void emitTypeDefList(ArrayRef<AttrOrTypeDef> defs);
+  /// Emit the code to dispatch between different defs during parsing/printing.
+  void emitParsePrintDispatch(ArrayRef<AttrOrTypeDef> defs);
+
+  /// The set of def records to emit.
+  std::vector<const llvm::Record *> defRecords;
+  /// The attribute or type class to emit.
+  /// The stream to emit to.
+  raw_ostream &os;
+  /// The prefix of the tablegen def name, e.g. Attr or Type.
+  StringRef defType;
+  /// The C++ base value type of the def, e.g. Attribute or Type.
+  StringRef valueType;
+  /// Flag indicating if this generator is for Attributes. False if the
+  /// generator is for types.
+  bool isAttrGenerator;
+};
+
+/// A specialized generator for AttrDefs.
+struct AttrDefGenerator : public DefGenerator {
+  AttrDefGenerator(const llvm::RecordKeeper &records, raw_ostream &os)
+      : DefGenerator(records.getAllDerivedDefinitionsIfDefined("AttrDef"), os,
+                     "Attr", "Attribute", /*isAttrGenerator=*/true) {}
+};
+/// A specialized generator for TypeDefs.
+struct TypeDefGenerator : public DefGenerator {
+  TypeDefGenerator(const llvm::RecordKeeper &records, raw_ostream &os)
+      : DefGenerator(records.getAllDerivedDefinitionsIfDefined("TypeDef"), os,
+                     "Type", "Type", /*isAttrGenerator=*/false) {}
+};
+
+void emitTypeConstraintDecls(const llvm::RecordKeeper &records,
+                             raw_ostream &os);
+
+void emitTypeConstraintDefs(const llvm::RecordKeeper &records, raw_ostream &os);
 
 } // namespace tblgen
 } // namespace mlir
